fixed messaginginapp tests
